Scripts/HotGameScript/requirements.txt
# 移动端自动化
Appium-Python-Client>=3.0.0
selenium>=4.15.0

# OCR识别
paddleocr>=2.7.0
paddlepaddle>=2.5.0
# 或使用 pytesseract
# pytesseract>=0.3.10

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0

# 网页自动化和请求
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 数据处理
pandas>=2.0.0

# 异步支持
aiohttp>=3.9.0
aiofiles>=23.2.0

# 工具库
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.66.0

# Chrome DevTools Protocol
pychrome>=0.2.4

# 自然语言处理（可选，用于更精确的游戏名称提取）
# openai>=1.0.0  # 如果使用OpenAI API
jieba>=0.42.1  # 中文分词
